app/reranker.py
import os
import json
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

class Reranker:
    """
    Stub for LLM-based reranking via Nebius AI Studio.
    """
    def __init__(self, model: str = None):
        base_url = os.getenv("NEBIUS_BASE_URL", "https://api.studio.nebius.ai/v1/")
        api_key = os.getenv("NEBIUS_API_KEY") or os.getenv("OPENAI_API_KEY")
        if not api_key:
            raise RuntimeError("Set NEBIUS_API_KEY or OPENAI_API_KEY environment variable")
        self.client = OpenAI(base_url=base_url, api_key=api_key)
        self.model = model or os.getenv("NEBIUS_MODEL", "meta-llama/Meta-Llama-3.1-405B-Instruct")

    def rerank(self, facet_scores, texts):
        """
        Re-rank top-k facet candidates using LLM.
        Currently returns input unmodified.
        """
        # facet_scores: list of (facet_id, score)
        # texts: list of preprocessed utterance strings
        # TODO: construct prompt with facets and invoke self.client.chat.completions.create
        return facet_scores
