requirements.txt
flask
numpy
torch
pytorch-lightning
onnxruntime
pyliwc
openai
python-dotenv
requests
scikit-learn
transformers
sentencepiece
sentence-transformers
